Revert "pinebook: Bump the kernel version for the dkms build"
This reverts commit 9db8801766828d007feaf029e0a06262111e3276.

This is annoying - need to come up with a better way to build the kernel
module/dkms in the setup.  In the 2020.3 release, we have the kali1
kernel but in kali-rolling, we've bumped to kali3 so I need to downgrade
the kernel and rebuild the release image before reverting this revert.
It's ugly but the server won't let me force push so it is what it is.